pyproject.toml
[tool.poetry]
name = "wnba-fantasy-league"
version = "0.1.0"
description = "A lightweight fantasy women's basketball web app for a season-long fantasy competition"
authors = ["Grant Harris <me@grantharris.tech>"]

[[tool.poetry.packages]]
include = "app"

[tool.poetry.dependencies]
python = "^3.10"
fastapi = "^0.101.0"
uvicorn = {extras = ["standard"], version = "^0.23.0"}
sqlalchemy = "^2.0.39"
apscheduler = "^3.11.0"
passlib = "^1.7.4"
bcrypt = "^4.3.0"
python-dotenv = "^1.0.0"
httpx = "^0.28.1"
requests = "^2.32.3"
alembic = "^1.15.2"
tenacity = "^9.1.2"
pytest-httpx = "^0.35.0"
python-jose = {extras = ["cryptography"], version = "^3.4.0"}
python-multipart = "^0.0.20"
email-validator = "^2.2.0"
freezegun = "^1.5.1"

[tool.poetry.dev-dependencies]
black = {version = "^23.9.1", allow-prereleases = false}
ruff = "^0.0.279"
isort = "^5.12.0"
pytest = "^8.0.0"
pytest-asyncio = "^0.21.0"
pre-commit = "^4.2.0"
httpx = "^0.28.1"
pytest-httpx = "^0.35.0"

[tool.poetry.group.dev.dependencies]
ipykernel = "^6.29.5"
freezegun = "^1.5.1"

[build-system]
requires = ["poetry-core>=1.6.0"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 120
skip-string-normalization = true
skip-magic-trailing-comma = true

[tool.isort]
profile = "black"
line_length = 120
multi_line_output = 3

[tool.ruff]
line-length = 120
target-version = "py310"
# Only select error, fixable, and warning codes (skip docstring checks)
select = ["E", "F", "W", "C90"]
# Ignore whitespace, no-newline, line-length, and unused import warnings
ignore = ["E203", "W292", "E501", "F401"]

exclude = [
    ".git",
    ".venv",
    "__pycache__",
    "build",
    "dist",
]

[tool.ruff.per-file-ignores]
"app/services/roster.py" = ["C901"]
"app/services/bonus.py" = ["C901"]

[tool.pytest.ini_options]
minversion = "7.0"
addopts = "-ra -q"
testpaths = ["tests"]
markers = [
    "slow: tests that take a long time to run and can be skipped with --skip-slow"
]